requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
python-multipart==0.0.9
httpx==0.27.0
pytest==8.2.1
pytest-cov==5.0.0
alembic==1.13.1
aiosqlite>=0.19.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
flask_cors==4.0.0
structlog==24.1.0
redis==5.0.4
# Config and validation
pydantic==2.6.4
pydantic-settings==2.2.1
email-validator==2.1.0
requests==2.31.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
Pillow==10.3.0
# Pin all dependencies for reproducible builds
# Remove unused or duplicate packages before production
